+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## vX.X.X (XXXX-XX-XX)
 
 - Add `from_filename()` function.
+- Add `from_autodetect()` function.
 
 ## v0.0.1 (2014-07-15)